--- LA-MCTS/lamcts/turbo/turbo_1_test.py
@@ -1,355 +1,0 @@
-###############################################################################
-# Copyright (c) 2019 Uber Technologies, Inc.                                  #
-#                                                                             #
-# Licensed under the Uber Non-Commercial License (the "License");             #
-# you may not use this file except in compliance with the License.            #
-# You may obtain a copy of the License at the root directory of this project. #
-#                                                                             #
-# See the License for the specific language governing permissions and         #
-# limitations under the License.                                              #
-###############################################################################
-
-import math
-import sys
-from copy import deepcopy
-
-import gpytorch
-import numpy as np
-import torch
-from torch.quasirandom import SobolEngine
-
-from .gp import train_gp
-from .utils import from_unit_cube, latin_hypercube, to_unit_cube
-
-
-class Turbo1:
-    """The TuRBO-1 algorithm.
-
-    Parameters
-    ----------
-    f : function handle
-    lb : Lower variable bounds, numpy.array, shape (d,).
-    ub : Upper variable bounds, numpy.array, shape (d,).
-    n_init : Number of initial points (2*dim is recommended), int.
-    max_evals : Total evaluation budget, int.
-    batch_size : Number of points in each batch, int.
-    verbose : If you want to print information about the optimization progress, bool.
-    use_ard : If you want to use ARD for the GP kernel.
-    max_cholesky_size : Largest number of training points where we use Cholesky, int
-    n_training_steps : Number of training steps for learning the GP hypers, int
-    min_cuda : We use float64 on the CPU if we have this or fewer datapoints
-    device : Device to use for GP fitting ("cpu" or "cuda")
-    dtype : Dtype to use for GP fitting ("float32" or "float64")
-
-    Example usage:
-        turbo1 = Turbo1(f=f, lb=lb, ub=ub, n_init=n_init, max_evals=max_evals)
-        turbo1.optimize()  # Run optimization
-        X, fX = turbo1.X, turbo1.fX  # Evaluated points
-    """
-
-    def __init__(
-        self,
-        path,
-        X_init,
-        f,
-        lb,
-        ub,
-        n_init,
-        max_evals,
-        batch_size=1,
-        verbose=True,
-        use_ard=True,
-        max_cholesky_size=2000,
-        n_training_steps=50,
-        min_cuda=1024,
-        device="cpu",
-        dtype="float64"
-    ):
-
-        # Very basic input checks
-        assert lb.ndim == 1 and ub.ndim == 1
-        assert len(lb) == len(ub)
-        assert np.all(ub > lb)
-        assert max_evals > 0 and isinstance(max_evals, int)
-        assert n_init > 0 and isinstance(n_init, int)
-        assert batch_size > 0 and isinstance(batch_size, int)
-        assert isinstance(verbose, bool) and isinstance(use_ard, bool)
-        assert max_cholesky_size >= 0 and isinstance(batch_size, int)
-        assert n_training_steps >= 30 and isinstance(n_training_steps, int)
-        assert max_evals > n_init and max_evals > batch_size
-        assert device == "cpu" or device == "cuda"
-        assert dtype == "float32" or dtype == "float64"
-        if device == "cuda":
-            assert torch.cuda.is_available(), "can't use cuda if it's not available"
-
-        # Save function information
-        self.f = f
-        self.dim = len(lb)
-        self.lb = lb
-        self.ub = ub
-        self.path = path
-
-        # Settings
-        self.X_init = X_init
-        self.n_init = n_init
-        self.max_evals = max_evals
-        self.batch_size = batch_size
-        self.verbose = verbose
-        self.use_ard = use_ard
-        self.max_cholesky_size = max_cholesky_size
-        self.n_training_steps = n_training_steps
-
-        # Hyperparameters
-        self.mean = np.zeros((0, 1))
-        self.signal_var = np.zeros((0, 1))
-        self.noise_var = np.zeros((0, 1))
-        self.lengthscales = np.zeros((0, self.dim)) if self.use_ard else np.zeros((0, 1))
-
-        # Tolerances and counters
-        self.n_cand = min(100 * self.dim, 5000)
-        self.failtol = np.ceil(np.max([4.0 / batch_size, self.dim / batch_size]))
-        self.succtol = 3
-        self.n_evals = 0
-
-        # Trust region sizes
-        self.length_min = 0.5 ** 7
-        self.length_max = 1.6
-        self.length_init = 0.8
-
-        # Save the full history
-        self.X = np.zeros((0, self.dim))
-        self.fX = np.zeros((0, 1))
-
-        # Device and dtype for GPyTorch
-        self.min_cuda = min_cuda
-        self.dtype = torch.float32 if dtype == "float32" else torch.float64
-        self.device = torch.device("cuda") if device == "cuda" else torch.device("cpu")
-        if self.verbose:
-            print("Using dtype = %s \nUsing device = %s" % (self.dtype, self.device))
-            sys.stdout.flush()
-
-        # Initialize parameters
-        self._restart()
-
-    def _restart(self):
-        self._X = []
-        self._fX = []
-        self.failcount = 0
-        self.succcount = 0
-        self.length = self.length_init
-
-    def _adjust_length(self, fX_next):
-        if np.min(fX_next) < np.min(self._fX) - 1e-3 * math.fabs(np.min(self._fX)):
-            self.succcount += 1
-            self.failcount = 0
-        else:
-            self.succcount = 0
-            self.failcount += 1
-
-        if self.succcount == self.succtol:  # Expand trust region
-            self.length = min([2.0 * self.length, self.length_max])
-            self.succcount = 0
-        elif self.failcount == self.failtol:  # Shrink trust region
-            self.length /= 2.0
-            self.failcount = 0
-
-    def propose_rand_samples(self, nums_samples, lb, ub):
-        x = np.random.uniform(lb, ub, size = (nums_samples, self.dim) )
-        return x
-
-    def get_sample_ratio_in_region( self, cands, path ):
-        total = len(cands)
-        for node in path:
-            boundary = node[0].classifier.svm
-            if len(cands) == 0:
-                return 0, np.array([])
-            assert len(cands) > 0
-            cands = cands[ boundary.predict( cands ) == node[1] ] 
-            # node[1] store the direction to go
-        ratio = len(cands) / total
-        assert len(cands) <= total
-        return ratio, cands
-
-    def propose_rand_samples_sobol(self, nums_samples, path, lb, ub):
-        
-        #rejected sampling
-        selected_cands = np.zeros((1, self.dim))
-        seed   = np.random.randint(int(1e6))
-        sobol  = SobolEngine(dimension = self.dim, scramble=True, seed=seed)
-
-        ratio_check, centers = self.get_sample_ratio_in_region(self.X, path)
-        # no current samples located in the region
-        # should not happen
-        # print("ratio check:", ratio_check, len(self.X) )
-        # assert ratio_check > 0
-        if ratio_check == 0 or len(centers) == 0:
-            print('full rand')
-            return self.propose_rand_samples( nums_samples, lb, ub )
-        
-        lb_    = None
-        ub_    = None
-        
-        final_cands = []
-        for center in centers:
-            center = self.X[ np.random.randint( len(self.X) ) ]
-            cands  = sobol.draw(2000).to(dtype=torch.float64).cpu().detach().numpy()
-            ratio  = 1
-            L      = 0.0001
-            Blimit = np.max(ub - lb)
-            
-            while ratio == 1 and L < Blimit:                    
-                lb_    = np.clip( center - L/2, lb, ub )
-                ub_    = np.clip( center + L/2, lb, ub )
-                cands_ = deepcopy( cands )
-                cands_ = (ub_ - lb_)*cands_ + lb_
-                ratio, cands_ = self.get_sample_ratio_in_region(cands_, path)
-                if ratio < 1:
-                    final_cands.extend( cands_.tolist() )
-                L = L*2
-        final_cands      = np.array( final_cands )
-        if len(final_cands) > nums_samples:
-            final_cands_idx  = np.random.choice( len(final_cands), nums_samples )
-            return final_cands[final_cands_idx]
-        else:
-            if len(final_cands) == 0:
-                print('full rand')
-                return self.propose_rand_samples( nums_samples, lb, ub )
-            else:
-                return final_cands
-
-    def _create_candidates(self, X, fX, length, n_training_steps, hypers):
-        """Generate candidates assuming X has been scaled to [0,1]^d."""
-        # Pick the center as the point with the smallest function values
-        # NOTE: This may not be robust to noise, in which case the posterior mean of the GP can be used instead
-        assert X.min() >= 0.0 and X.max() <= 1.0
-
-        # Standardize function values.
-        mu, sigma = np.median(fX), fX.std()
-        sigma = 1.0 if sigma < 1e-6 else sigma
-        fX = (deepcopy(fX) - mu) / sigma
-
-        # Figure out what device we are running on
-        if len(X) < self.min_cuda:
-            device, dtype = torch.device("cpu"), torch.float64
-        else:
-            device, dtype = self.device, self.dtype
-
-        # We use CG + Lanczos for training if we have enough data
-        with gpytorch.settings.max_cholesky_size(self.max_cholesky_size):
-            X_torch = torch.tensor(X).to(device=device, dtype=dtype)
-            y_torch = torch.tensor(fX).to(device=device, dtype=dtype)
-            gp = train_gp(
-                train_x=X_torch, train_y=y_torch, use_ard=self.use_ard, num_steps=n_training_steps, hypers=hypers
-            )
-
-            # Save state dict
-            hypers = gp.state_dict()
-
-        # Create the trust region boundaries
-        x_center = X[fX.argmin().item(), :][None, :]
-        weights = gp.covar_module.base_kernel.lengthscale.cpu().detach().numpy().ravel()
-        weights = weights / weights.mean()  # This will make the next line more stable
-        weights = weights / np.prod(np.power(weights, 1.0 / len(weights)))  # We now have weights.prod() = 1
-        lb = np.clip(x_center - weights * length / 2.0, 0.0, 1.0)
-        ub = np.clip(x_center + weights * length / 2.0, 0.0, 1.0)
-
-        X_cand = self.propose_rand_samples_sobol(self.n_cand, self.path, lb, ub)
-
-        # Figure out what device we are running on
-        if len(X_cand) < self.min_cuda:
-            device, dtype = torch.device("cpu"), torch.float64
-        else:
-            device, dtype = self.device, self.dtype
-
-        # We may have to move the GP to a new device
-        gp = gp.to(dtype=dtype, device=device)
-
-        # We use Lanczos for sampling if we have enough data
-        with torch.no_grad(), gpytorch.settings.max_cholesky_size(self.max_cholesky_size):
-            X_cand_torch = torch.tensor(X_cand).to(device=device, dtype=dtype)
-            y_cand = gp.likelihood(gp(X_cand_torch)).sample(torch.Size([self.batch_size])).t().cpu().detach().numpy()
-
-        # Remove the torch variables
-        del X_torch, y_torch, X_cand_torch, gp
-
-        # De-standardize the sampled values
-        y_cand = mu + sigma * y_cand
-
-        return X_cand, y_cand, hypers
-
-    def _select_candidates(self, X_cand, y_cand):
-        """Select candidates."""
-        X_next = np.ones((self.batch_size, self.dim))
-        for i in range(self.batch_size):
-            # Pick the best point and make sure we never pick it again
-            indbest = np.argmin(y_cand[:, i])
-            X_next[i, :] = deepcopy(X_cand[indbest, :])
-            y_cand[indbest, :] = np.inf
-        return X_next
-
-    def optimize(self):
-        """Run the full optimization process."""
-        if len(self._fX) > 0 and self.verbose:
-            n_evals, fbest = self.n_evals, self._fX.min()
-            print(f"{n_evals}) Restarting with fbest = {fbest:.4}")
-            sys.stdout.flush()
-
-        # Initialize parameters
-        self._restart()
-        
-        # Generate and evalute initial design points
-        fX_init = np.array([[self.f(x)] for x in self.X_init])
-        
-        # Update budget and set as initial data for this TR
-        self.n_evals += self.n_init
-        self._X = deepcopy(self.X_init)
-        self._fX = deepcopy(fX_init)
-        
-        # Append data to the global history
-        self.X = np.vstack((self.X, deepcopy(self.X_init)))
-        self.fX = np.vstack((self.fX, deepcopy(fX_init)))
-        
-        if self.verbose:
-            fbest = self._fX.min()
-            print(f"Starting from fbest = {fbest:.4}")
-            sys.stdout.flush()
-        
-        # Thompson sample to get next suggestions
-        while self.n_evals < self.max_evals and self.length >= self.length_min:
-            
-            # Warp inputs
-            X = to_unit_cube(deepcopy(self._X), self.lb, self.ub)
-            
-            # Standardize values
-            fX = deepcopy(self._fX).ravel()
-            
-            # Create th next batch
-            X_cand, y_cand, _ = self._create_candidates(
-                X, fX, length=self.length, n_training_steps=self.n_training_steps, hypers={}
-            )
-            X_next = self._select_candidates(X_cand, y_cand)
-            
-            # Undo the warping
-            X_next = from_unit_cube(X_next, self.lb, self.ub)
-            
-            # Evaluate batch
-            fX_next = np.array([[self.f(x)] for x in X_next])
-            
-            # Update trust region
-            self._adjust_length(fX_next)
-            
-            # Update budget and append data
-            self.n_evals += self.batch_size
-            self._X = np.vstack((self._X, X_next))
-            self._fX = np.vstack((self._fX, fX_next))
-            
-            if self.verbose and fX_next.min() < self.fX.min():
-                n_evals, fbest = self.n_evals, fX_next.min()
-                print(f"{n_evals}) New best: {fbest:.4}")
-                sys.stdout.flush()
-            
-            # Append data to the global history
-            self.X = np.vstack((self.X, deepcopy(X_next)))
-            self.fX = np.vstack((self.fX, deepcopy(fX_next)))
-        
-        return self.X, self.fX
